Add files via upload
First draft, was able to pull in csv and create a dictionary for the first round match-ups. Think a dictionary is going to be the best way to go for specifying match-ups, so that way in future round, the values can be a list of every possible team that team could play in that round.