Add files via upload
Code for 10 best brackets, implemented with a for loop. One hiccup I can't figure out: the 6th best bracket (97.306) has a better objective value than the 5th (97.301). The 5th should be better because that's how maximization problems work, right? Maybe it's the perfect storm of constraints to where the objective value is able to jump up higher than it's previous objective value?